Extract FairRoundRobinDispatcher from PacketDispatcher
